move routers
