+++ site/app/libraries/routers/AnnotatedRouteLoader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\libraries;
+namespace app\libraries\routers;
 
 use Symfony\Component\Routing\Loader\AnnotationClassLoader;
 use Symfony\Component\Routing\Route;
