+++ site/app/libraries/routers/ApiRouter.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace app\libraries;
+namespace app\libraries\routers;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\HttpFoundation\Request;
@@ -11,6 +11,7 @@ use Symfony\Component\Routing\Exception\ResourceNotFoundException;
 use Symfony\Component\Routing\Loader\AnnotationDirectoryLoader;
 use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\Common\Annotations\AnnotationException;
+use app\libraries\Core;
 
 
 class ApiRouter {
